+++ src/App.js
@@ -1,8 +1,8 @@
 import logo from './logo.svg';
 import './App.css';
-import Header from './components/Header'
-import Nav from './components/Navbar';
-import Profile from './components/profile';
+import Header from './components/header/Header'
+import Nav from './components/navbar/Navbar';
+import Profile from './components/profile/profile';
