+++ src/components/profile/profile.module.css
@@ -1,0 +1,47 @@
+.content {
+    grid-area: c;
+    background-color: rgb(0, 0, 0);
+    padding: 10px;
+}
+
+.infoUser {
+    display: flex;
+    gap: 25px;
+}
+
+.imageTop {
+    width: 100%;
+    max-width: 1001px;
+    height: 200px;
+    object-fit: cover;
+    background-image: url(https://static.vecteezy.com/system/resources/thumbnails/012/979/480/small/black-and-blue-abstract-banner-with-polygonal-pattern-template-for-your-header-and-presentation-modern-background-vector.jpg);
+    background-position: left;
+    background-repeat: no-repeat;
+    background-size: cover;
+    margin-bottom: 20px;
+}
+
+.avatar {
+    width: 200px;
+    height: 200px;
+}
+
+.avatar img {
+    width: 100%;
+    border-radius: 5px;
+}
+
+.description {
+    display: flex;
+    flex-direction: column;
+    gap: 15px;
+}
+
+.description div:first-of-type {
+    font-size: 25px;
+}
+
+.description div {
+    color: #47525e;
+    font-size: 20px;
+}
